# openapi.yaml
openapi: 3.0.0
info:
  title: Donor's Choice API
  description: Donor's Choice Project Application API
  version: 0.0.1

servers:
  - url: 'http://localhost:3000/'
    description: Localhost running on your computer
  - url: 'https://arcane-escarpment-54741.herokuapp.com/'
    description: Heroku site

tags:
  - name: user
    description: Operations about users
  - name: password
    description: Operations about password changing
  - name: charity
    description: Operations about charities
  - name: pickup
    description: Operations about pickups
  - name: donation
    description: Operations about donations
  - name: contact
    description: Operations about contacts

paths:

# /users/signup POST configuration:
# No parameters are configured for a POST
  /users/signup:
    post:
      tags:
        - "user"
      summary: Sign up a new user to the system
      description: Signup to the system
      requestBody:
        description: User information required to create an account
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: Richard
                lastname:
                  type: string
                  example: Polisuk
                username:
                  type: string
                  example: email@example.com
                password:
                  type: string
                  example: password
      responses:
        '200':
          description: User is logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Registration Successful!
        '422':
          description: Error - Unprocessable Entity

# /users/verify GET configuration:
  /users/verify:
    get:
      tags:
        - "user"
      summary: Verifies the user to the system
      description: Verify the user to the system
      parameters:
        - name: email
          in: query
          schema:
            type: string
            example: polisuk@gmail.com
        - name: code
          in: query
          schema:
            type: string
            example: 63d840b9b19b722fbb0ba1eca846db88
      responses:
        '301':
          description: Error - Missing information, re-directed
        '422':
          description: Error - Unprocessable Entity

# /users/login POST configuration:
  /users/login:
    post:
      tags:
        - "user"
      summary: Logs user into the system
      description: Log in to the system with a valid userid
      requestBody:
        description: User that is allowed to log in
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginSend'
        required: false
      responses:
        '200':
          description: User is logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginReceive'
        '401':
          description: User is not authorized

# /users/forgot POST configuration:
# No parameters are configured for a POST
  /users/forgot:
    post:
      tags:
        - "password"
      summary: Reset a forgotten password
      description: Resets a forgotten password
      requestBody:
        description: User information required to reset the password
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: email@example.com
      responses:
        '200':
          description: Request to reset password was succesfully processed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Request to reset password was succesfully processed.
        '422':
          description: Error - No account with that username exists.

# /users/verify GET configuration:
  /users/get:
    get:
      tags:
        - "password"
      summary: Gets the reset tokens for a user
      description: Determines if a password reset has been requested
      parameters:
        - name: username
          in: query
          schema:
            type: string
            example: polisuk@gmail.com
      responses:
        '200':
          description: User is logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  resetPasswordToken:
                    type: string
                    example: 3e1a05b035c13f0a7663d71127712463176deb50
                  resetPasswordExpires:
                    type: string
                    example: 2021-06-16T17:37:27.944Z
        '422':
          description: Error - Unprocessable Entity

# /users/forgot POST configuration:
# No parameters are configured for a POST
  /users/reset/{token}:
    put:
      tags:
        - "password"
      summary: Reset a forgotten password
      description: Resets a forgotten password
      parameters:
      - in: path
        name: token
        schema:
          type: string
          example: 7e14038bfa597141b7dcd8f51acd43f4c8e54730
        required: true
      requestBody:
        description: User information required to reset the password
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: password
      responses:
        '200':
          description: Request to reset password was succesfully processed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Request to reset password was succesfully processed.
        '422':
          description: Error - No account with that username exists.

# /{userId}/change-password PUT configuration: 
  '/users/{username}/change-password/':
    put:
      tags:
        - "password"
      security:
        - bearerAuth: []
      summary: Change the password for the user
      parameters:
        - name: username
          in: path
          description: Email of the user that we need to change the password for.
          required: true
          schema:
            type: string
            example: polisuk@gmail.com
      requestBody:
        description: User with new values of properties
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: password
                new_password:
                  type: string
                  example: password
      responses:
        '200':
          description: User is deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: User is not authorized
        '404':
          description: Argument passed in must be a single String of 12 bytes or a string of 24 hex characters OR password or username is incorrect

# /charities GET configuration: 
  /charities:
    get:
      tags:
        - "charity"
      summary: Get the charity information
      description: Return all the details charity information.
      parameters:
        - in: query
          name: legalname
          schema:
            type: string
            example: Seneca
          required: true
        - in: query
          name: page
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Your request charities have been successfully returned.
          content:
            application/json:
              schema:
                properties:
                  charities: 
                    type: array
                    items: 
                        type: object
                        properties: 
                          businessnumber: 
                            type: string
                            example: 107967168RR0001
                          legalname: 
                            type: string
                            example: SENECA COLLEGE OF APPLIED ARTS & TECHNOLOGY
                          addressline1: 
                            type: string
                            example: NAAZ MOHAMED DIRECTOR FINANCE
                          addressline2: 
                            type: string
                            example: 1750 FINCH AVENUE EAST
                          city: 
                            type: string
                            example: TORONTO
                          province: 
                            type: string
                            example: ON
                          postalcode: 
                            type: string
                            example: M2J2X5
                          phone: 
                            type: string
                            example: (416) 491-5050
                          _id: 
                            type: string
                            example: 6092d2ba9005b33f68ec5a67
                  totalPages: 
                    type: integer
                    format: int32
                    example: 1
                  currentPage: 
                    type: string
                    example: 1
#
# /charities/byBN GET configuration: 
  /charities/byBN:
    get:
      tags:
        - "charity"
      summary: Get the charity information for a specific charity number
      description: Return the details of the specific charity.
      parameters:
        - in: query
          name: businessnumber
          schema:
            type: string
            example: 107967168RR0001
          required: true
      responses:
        '200':
          description: Your request charity have been successfully returned.
          content:
            application/json:
              schema:
                properties:
                  found: 
                    type: boolean
                    example: true
                  charity: 
                    type: object
                    properties: 
                      businessnumber: 
                        type: string
                        example: 107967168RR0001
                      legalname: 
                        type: string
                        example: SENECA COLLEGE OF APPLIED ARTS & TECHNOLOGY
                      addressline1: 
                        type: string
                        example: NAAZ MOHAMED DIRECTOR FINANCE
                      addressline2: 
                        type: string
                        example: 1750 FINCH AVENUE EAST
                      city: 
                        type: string
                        example: TORONTO
                      province: 
                        type: string
                        example: ON
                      postalcode: 
                        type: string
                        example: M2J2X5
                      phone: 
                        type: string
                        example: (416) 491-5050
                      _id: 
                        type: string
                        example: 6092d2ba9005b33f68ec5a67
        '422':
          description: Charity was not found or invalid format.
#
#
# /pickups/schedule POST configuration:
# No parameters are configured for a POST
  /pickups/schedule:
    post:
      tags:
        - "pickup"
      security:
        - bearerAuth: []
      summary: Schedule a pickup
      description: Schedule a pickup
      requestBody:
        description: User information required to schedule a pickup
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: polisuk@gmail.com
                contactname:
                  type: string
                  example: Sam Example
                address:
                  type: string
                  example: 123 Example Road
                city:
                  type: string
                  example: Toronto
                province:
                  type: string
                  example: Ontario
                postalcode:
                  type: string
                  example: H0H 0H0
                phone:
                  type: string
                  example: 555-123-4567
                items:
                  type: array
                  items:
                    type: string
                    example: MenClothes, WomenClothes
                donations:
                  type: array
                  items:
                    type: object
                    properties:
                      businessnumber:
                        type: string
                        example: 821832037RR0001
                      percent:
                        type: string
                        example: 50
                  example:
                    - businessnumber: 821832037RR0001
                      percent: 50
                    - businessnumber: 135518140RR0001
                      percent: 50
                pickupdate:
                  type: string
                  example: 2021-12-01
                pickuptime:
                  type: string
                  enum: [morning, afternooon, evening]
                  example: morning
      responses:
        '200':
          description: User is logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Registration Successful!
        '422':
          description: Error - Unprocessable Entity
#
# /{userId}/change-password PUT configuration: 
  '/pickups/update/{pickupId}/':
    put:
      tags:
        - "pickup"
      security:
        - bearerAuth: []
      summary: Updates the pickup record
      parameters:
        - name: pickupId
          in: path
          description: Object ID of the pickup that we need to update.
          required: true
          schema:
            type: string
            example: 60d1f274f8bef37d84298e23
      requestBody:
        description: User with new values of properties
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: cancelled
      responses:
        '200':
          description: Record is updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Updated successfully.
        '401':
          description: User is not authorized
        '404':
          description: Argument passed in must be a single String of 12 bytes or a string of 24 hex characters OR password or username is incorrect
#
# /pickups/requests GET configuration:
  /pickups/requests:
    get:
      tags:
        - "pickup"
      security:
        - bearerAuth: []
      summary: Gets all the scheduled pickups 
      description:  Gets all the scheduled pickups. **This operation requires admin role.**
      responses:
        '200':
          description: Array of donations
        '422':
          description: No donations found. or Invalid format.
#
# /donations/create POST configuration:
# No parameters are configured for a POST
  /donations/create:
    post:
      tags:
        - "donation"
      security:
        - bearerAuth: []
      summary: Create a donation
      description: Information about a donation. **This operation requires admin role.**
      requestBody:
        description: User information required to create the donation
        content:
          application/json:
            schema:
              type: object
              properties:
                pickupid:
                  type: string
                  example: 60d1f274f8bef37d84298e23
                donorid:
                  type: string
                  example: polisuk@gmail.com
                adminid:
                  type: string
                  example: polisuk@gmail.com
                businessnumber:
                  type: string
                  example: 821832037RR0001
                amount:
                  type: number
                  example: 50
      responses:
        '200':
          description: Request to reset password was succesfully processed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Successfully added.
#
# /donations/history GET configuration:
  /donations/history:
    get:
      tags:
        - "donation"
      summary: Gets all the donations made on behalf of the user
      description: Gets all the donations made on behalf of the user. **This operation requires admin role.**
      parameters:
        - name: donorid
          in: query
          schema:
            type: string
            example: polisuk@gmail.com
      responses:
        '200':
          description: Array of donations
        '422':
          description: No donations found. or Invalid format.
#
# /contacts/create POST configuration:
# No parameters are configured for a POST
  /contacts/create:
    post:
      tags:
        - "contact"
      summary: Create a new contact entry
      description: Allows the user to create a contact
      requestBody:
        description: User information required to create an contact
        content:
          application/json:
            schema:
              type: object
              properties:
                fname:
                  type: string
                  example: Richard
                lname:
                  type: string
                  example: Polisuk
                subject:
                  type: string
                  example: Test subject
                email:
                  type: string
                  example: sample@example.com
                message:
                  type: string
                  example: This is a test message. This is a second sentence.
      responses:
        '200':
          description: Contact request is logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  status:
                    type: string
                    example: Successfully added.
        '422':
          description: Error - Unprocessable Entity
#
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    # Schema for input of /users/login
    UserLoginSend:
      required:
        - username
        - password
      type: object
      properties:
        username:
          type: string
          example: email@example.com
        password:
          type: string
          example: password
    # Schema for output of /users/login
    UserLoginReceive:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        _id:
          type: string
        verfiied:
          type: boolean
        status:
          type: string
    # Schema for output of /users
    UserGetReceive:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        admin:
          type: boolean
        _id:
          type: string
        username:
          type: string
        __v:
          type: integer
          
    User:
      required:
        - _id
        - companies
        - name
      type: object
      properties:
        _id:
          type: integer
        isPublic:
          type: boolean
        name:
          type: string
        books:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
        companies:
          type: array
          items:
            type: string
    updateUser:
      required:
        - companies
        - name
      type: object
      properties:
        isPublic:
          type: boolean
        name:
          type: string
        books:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
        companies:
          type: array
          items:
            type: string